Feat[#250] : Weekly Weather
